+++ src/marshmallow/fields.py
@@ -11,7 +11,7 @@ import typing
 import warnings
 from collections.abc import Mapping as _Mapping
 
-from marshmallow import validate, utils, class_registry
+from marshmallow import validate, utils, class_registry, types
 from marshmallow.base import FieldABC, SchemaABC
 from marshmallow.utils import (
     is_collection,
@@ -328,7 +328,7 @@ class Field(FieldABC):
         self,
         value: typing.Any,
         attr: str = None,
-        data: typing.Dict[str, typing.Any] = None,
+        data: typing.Mapping[str, typing.Any] = None,
         **kwargs
     ):
         """Deserialize ``value``.
@@ -388,7 +388,7 @@ class Field(FieldABC):
         self,
         value: typing.Any,
         attr: typing.Optional[str],
-        data: typing.Optional[typing.Dict[str, typing.Any]],
+        data: typing.Optional[typing.Mapping[str, typing.Any]],
         **kwargs
     ):
         """Deserialize value. Concrete :class:`Field` classes should implement this method.
@@ -480,13 +480,11 @@ class Nested(Field):
 
     def __init__(
         self,
-        nested: typing.Union[
-            SchemaABC, typing.Type[SchemaABC], str, typing.Callable[[], SchemaABC]
-        ],
+        nested: typing.Union[SchemaABC, type, str, typing.Callable[[], SchemaABC]],
         *,
         default: typing.Any = missing_,
-        only: typing.Union[typing.Sequence, typing.Set] = None,
-        exclude: typing.Union[typing.Sequence, typing.Set] = tuple(),
+        only: types.StrSequenceOrSet = None,
+        exclude: types.StrSequenceOrSet = (),
         many: bool = False,
         unknown: str = None,
         **kwargs
@@ -683,9 +681,7 @@ class List(Field):
 
     default_error_messages = {"invalid": "Not a valid list."}
 
-    def __init__(
-        self, cls_or_instance: typing.Union[Field, typing.Type[Field]], **kwargs
-    ):
+    def __init__(self, cls_or_instance: typing.Union[Field, type], **kwargs):
         super().__init__(**kwargs)
         try:
             self.inner = resolve_field_instance(cls_or_instance)
@@ -1446,8 +1442,8 @@ class Mapping(Field):
 
     def __init__(
         self,
-        keys: typing.Union[Field, typing.Type[Field]] = None,
-        values: typing.Union[Field, typing.Type[Field]] = None,
+        keys: typing.Union[Field, type] = None,
+        values: typing.Union[Field, type] = None,
         **kwargs
     ):
         super().__init__(**kwargs)
@@ -1592,7 +1588,7 @@ class Url(String):
         self,
         *,
         relative: bool = False,
-        schemes: typing.Union[typing.Sequence, typing.Set] = None,
+        schemes: types.StrSequenceOrSet = None,
         require_tld: bool = True,
         **kwargs
     ):